Simulating enrollment dates

This function simulates enrollment dates using either poisson or binomial distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   0   1   4   5   6   8  10  10  10  11  11  12  14  17  20  21  21
#>  [19]  22  24  25  27  27  30  31  32  33  33  34  36  37  39  41  42  45  48
#>  [37]  48  48  50  50  51  53  53  54  57  59  65  66  66  68  69  70  73  74
#>  [55]  75  77  78  78  81  81  85  90  90  94  96  99 103 104 105 107 109 109
#>  [73] 110 121 129 129 129 129 129 131 132 132 136 138 139 140 142 142 144 145
#>  [91] 146 147 148 148 148 149 151 152 153 153

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   8  10  15  15  15  17  18  25  25  26  27  28  29  30  30  30  30
#>  [19]  31  32  33  34  34  34  35  36  38  40  41  43  43  44  45  45  45  46
#>  [37]  46  46  47  49  49  49  50  51  53  53  53  53  54  54  54  55  55  56
#>  [55]  57  57  59  60  60  61  62  62  63  63  64  64  65  65  65  65  66  67
#>  [73]  67  67  68  69  69  69  69  69  70  70  71  71  71  71  72  72  72  73
#>  [91]  73  73  74  74  74  74  75  75  76  77  77  77  77  77  77  78  78  79
#> [109]  79  79  80  80  80  81  81  81  83  83  83  83  83  83  84  85  87  87
#> [127]  87  87  88  88  88  88  88  89  89  90  90  90  91  91  92  93  93  93
#> [145]  93  94  94  94  94  94  94  95  95  96  96  96  96  97  97  98  98  98
#> [163]  98  99  99  99  99 100 100 100 101 101 102 102 102 102 102 103 103 104
#> [181] 104 104 104 105 106 106 106 107 108 109 110 110 111 111 111 111 112 112
#> [199] 113 113