Normal distribution for Bayesian Adaptive Trial

Simulation for normal distribution for Bayesian Adaptive trial with different inputs to control for power, sample size, type 1 error rate, and etc.

normalBACT(mu_control, sd_control, mu_treatment, sd_treatment, N_total, lambda,
  lambda_time, interim_look, EndofStudy, prior = c(1, 1), block = 2,
  rand_ratio = c(1, 1), prop_loss_to_followup = 0.15, h0 = 0,
  futility_prob = 0.05, expected_success_prob = 0.9, prob_ha = 0.95,
  N_impute = 100, number_mcmc = 1000)

Arguments

mu_control
scalar. Mean of the control group.

sd_control
scalar. Standard deviation of the control group.

mu_treatment
scalar. Mean of the treatment group.

sd_treatment
scalar. Standard deviation of the treatment group.

N_total
scalar. Total sample size

lambda
vector. Lambda for different enrollment rates across times.

lambda_time
vector. Same size as lambda, denote times at lambda changes.

interim_look
vector. Sample size for interim looks, total size not included.

EndofStudy
scalar. Length of the study.

prior
vector. Prior value of beta rate, beta(a0, b0)

block
scalar. Block size for randomization to be implemented.

rand_ratio
vector. Randomization ratio for control to treatment. Integer values mapping the size of the block.

prop_loss_to_followup
scalar. Proportion of loss in follow up.

h0
scalar. treshold for comparing two proportions. default at 0.

futility_prob
scalar. Type 1 error rate.

expected_success_prob
scalar. Power of the study.

prob_ha
scalar. Probability of alternative hypothesis.

N_impute
scalar. Number of imputation for monte carlo simulation for missing data.

number_mcmc
scalar. Number of Monte Carlo Markov Chain in sampling posterior.

Value

a list of output

Examples

normalBACT(mu_control = 10, mu_treatment = 8,
           sd_control = 0.8, sd_treatment = 1.2, N_total = 300,
           lambda = c(0.3, 1), lambda_time = c(25),
           interim_look = c(110, 140, 220, 270),
           EndofStudy = 50)

#> [1] 110
#> [1] 110

#> $prob_ha
#> [1] 0.95
#> 
#> $N_treatment
#> [1] 48
#> 
#> $N_control
#> [1] 47
#> 
#> $N_total
#> [1] 110
#> 
#> $N_max
#> [1] 300
#> 
#> $stop_futility
#> [1] 1
#> 
#> $stop_expected_success
#> [1] 0
#> 
#> $prob_ha_final
#> [1] 0.873
#> 
#> $est_final
#> [1] -1.813827
#> 
#> $est_int
#> [1] -2.170093
#> 
#> $MLE_est
#> treatment 
#> -1.834474 
#> 
#> $MLE_est_int
#> treatment 
#> -2.212571 
#> 
#> $mu_treatment
#> [1] 8
#> 
#> $mu_control
#> [1] 10
#> 
#> $sd_treatment
#> [1] 1.2
#> 
#> $sd_control
#> [1] 0.8
#>