Randomization scheme

Implements randomization scheme for control and treatment with different randomization ratios and block

randomization(N_total, block = 2, scheme = c(1, 1))

Arguments

N_total
a numeric value of total sample size for randomization scheme

block
a integer value of the block size for randomization, it needs to be a multiple of the sum of scheme

scheme
a numeric vector of the randomization scheme c(control, treatment)

Value

the randomization scheme with 0, 1 for control and treatment

Examples

#implementing randomization scheme for control to treatment with 1:1.5 randomization ratio
randomization(100, block = 5, c(2, 3))

#>   [1] 0 0 1 1 1 1 1 0 1 0 1 1 0 0 1 1 0 1 1 0 0 1 1 1 0 1 0 1 1 0 1 0 1 0 1 0 0
#>  [38] 1 1 1 0 1 1 0 1 1 1 1 0 0 1 1 0 1 0 1 1 0 1 0 1 0 1 0 1 0 1 1 0 1 0 0 1 1
#>  [75] 1 0 1 1 1 0 0 1 1 1 0 0 1 1 0 1 0 1 0 1 1 1 0 0 1 1

#randomization scheme with 2:1 for control to treatment
randomization(70, block = 9, c(2, 1))

#>  [1] 0 0 0 1 1 0 0 0 1 0 0 1 0 0 1 0 1 0 0 0 0 0 1 1 0 0 1 0 1 0 1 0 1 0 0 0 0 1
#> [39] 0 0 1 0 0 1 0 1 1 1 0 0 0 0 0 0 0 1 0 1 0 0 0 1 0 1 0 0 0 0 1 0