Simulating enrollment dates

This function simulates enrollment dates using either poisson or binomial distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   1   2   2   3   3   5  10  11  12  14  14  16  19  19  21  22  23
#>  [19]  28  28  28  29  34  41  42  44  45  45  51  51  52  53  53  54  56  56
#>  [37]  57  60  61  62  62  63  64  65  67  67  67  67  69  69  71  71  73  73
#>  [55]  75  77  78  81  81  84  84  85  86  88  89  94  94 103 103 104 104 105
#>  [73] 106 108 109 109 110 110 112 112 113 113 114 114 115 118 118 119 120 120
#>  [91] 122 127 128 129 133 134 135 135 136 137

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   4   4   4   8  12  16  18  20  20  24  27  29  29  31  31  33  35
#>  [19]  35  35  37  37  37  37  39  40  42  42  45  45  46  49  49  50  51  51
#>  [37]  52  53  53  54  54  55  55  55  55  56  56  57  58  58  59  60  61  62
#>  [55]  62  62  63  63  65  65  65  65  66  67  67  67  67  68  68  69  69  69
#>  [73]  71  71  72  73  73  73  73  74  74  75  76  77  77  78  78  78  80  80
#>  [91]  80  81  81  81  81  82  82  83  84  84  85  86  86  87  87  87  88  89
#> [109]  89  90  90  90  91  92  93  93  93  93  96  96  96  96  97  97  97  97
#> [127]  99 100 101 102 102 102 102 102 103 103 103 103 103 104 104 105 105 106
#> [145] 106 107 107 108 108 109 109 109 109 109 109 110 110 111 111 112 112 112
#> [163] 113 113 113 114 114 115 116 116 116 116 116 117 118 118 119 119 119 120
#> [181] 121 121 121 122 124 124 124 125 126 126 126 127 127 127 127 127 128 128
#> [199] 128 129