Simulating enrollment dates

This function simulates enrollment dates using either poisson or binomial distribution

enrollment(param, N_total, time = NULL)

Arguments

param
a vector of lambda in poisson

N_total
a numeric value of total sample size

time
a vector of the length(param) - 1 indicating end of time when a specific lambda is used

Value

a vector of enrollment dates

Examples

enrollment(param = c(0.003, 0.7), 100, time = 10)

#>   [1]   0   1   3   4   4   6   8  10  10  11  14  15  16  17  18  21  21  26
#>  [19]  31  32  33  34  34  35  35  37  37  37  41  42  43  43  44  46  46  52
#>  [37]  54  56  57  58  60  61  61  62  66  66  67  68  69  72  72  74  75  89
#>  [55]  91  98 101 103 103 104 104 105 105 107 109 110 110 111 112 115 116 117
#>  [73] 118 119 120 129 130 131 133 133 134 135 135 135 135 136 138 138 138 139
#>  [91] 141 143 143 144 145 145 146 147 148 149

enrollment(param = c(0.3, 0.5, 0.9, 1.2, 2.1), 200, c(20, 30, 40, 60))

#>   [1]   0   1   4   8   9  14  15  16  19  19  19  21  23  28  28  29  30  30
#>  [19]  33  34  34  35  35  37  38  38  38  39  41  41  41  42  42  42  46  46
#>  [37]  46  47  50  51  51  52  54  54  56  56  56  58  58  58  58  59  60  60
#>  [55]  60  60  61  61  61  62  62  62  63  63  63  64  64  64  65  65  66  66
#>  [73]  67  67  67  68  70  70  70  71  71  71  71  72  72  72  73  74  75  75
#>  [91]  76  76  77  77  78  78  78  78  79  80  81  81  81  81  81  82  83  83
#> [109]  84  84  86  87  87  89  89  89  90  90  91  93  93  94  95  96  96  97
#> [127]  97  97 100 100 100 101 102 102 103 103 104 106 106 106 106 107 107 108
#> [145] 109 109 110 110 110 111 111 111 112 112 112 113 113 113 113 113 114 114
#> [163] 115 115 116 116 117 118 118 118 119 120 120 120 122 122 122 123 123 125
#> [181] 125 126 126 127 128 128 129 129 130 131 131 131 131 132 132 132 133 133
#> [199] 133 133